cefr-report/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.2.0
python-dotenv==1.0.1
